tests: Fix nondeterministic failures in is_full()
The test suite for the following functions exhibited nondeterministic
behavior:

- {Reader, Writer}::len()
- {Reader, Writer}::is_full()
- {Reader, Writer}::len()

This is because they used the functions "poll_fill_bytes" and
"poll_drain_bytes", which introduce a random chance to yield in order to
improve contention under certain conditions. However this yield has a
chance of kicking in and making these functions read/write 0 bytes
instead of the desired number of bytes, causing tests to fail.

To fix this issue I have replaced the functions used in these tests with
"try_fill" and "try_drain". These do not yield and therefore allow these
tests to be deterministic.

Closes #18

Signed-off-by: John Nunley <dev@notgull.net>