feat: Implement AsyncBufRead and related methods on Reader and Writer for directly accessing the buffer
AsyncBufRead is useful for anything parsing the received data to avoid another copy to an intermediate buffer.

There is no standard trait for the Writer side, but my use case is for making multiple readers at different positions over the same file descriptor using read_at to virtualize the file cursor. The existing API can only pass a buffer to read and not read_at.